+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+No changes so far.
+
+## [v0.4.3] - February 16, 2020
+
 Fixed:
 
 * windows showstopper: permissions checking always failed. Removed check on windows platform.
@@ -97,6 +101,7 @@ Added:
 * secret key distribution using OpenPGP keys
 
 [Unreleased]: https://github.com/julian7/redact
+[v0.4.3]: https://github.com/julian7/redact/releases/tag/v0.4.3
 [v0.4.2]: https://github.com/julian7/redact/releases/tag/v0.4.2
 [v0.4.1]: https://github.com/julian7/redact/releases/tag/v0.4.1
 [v0.4.0]: https://github.com/julian7/redact/releases/tag/v0.4.0